backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
